errors.go
package lugo

import (
	"fmt"
	"strconv"
	"strings"

	lua "github.com/yuin/gopher-lua"
)

// LuaStackTrace represents a Lua stack frame
type LuaStackTrace struct {
	Source   string // Source file or chunk name
	Line     int    // Line number
	Function string // Function name
}

// LuaError represents a Lua-specific error with stack trace
type LuaError struct {
	BaseError *Error // Renamed from Error to BaseError to avoid naming conflict
	Stack     []LuaStackTrace
}

// Error implements the error interface
func (e *LuaError) Error() string {
	var b strings.Builder
	b.WriteString(e.BaseError.Message)
	if e.BaseError.Cause != nil {
		b.WriteString(": ")
		b.WriteString(e.BaseError.Cause.Error())
	}
	if len(e.Stack) > 0 {
		b.WriteString("\nLua Stack Trace:")
		for _, frame := range e.Stack {
			if frame.Function != "" {
				b.WriteString(fmt.Sprintf("\n  at %s (%s:%d)", frame.Function, frame.Source, frame.Line))
			} else {
				b.WriteString(fmt.Sprintf("\n  at %s:%d", frame.Source, frame.Line))
			}
		}
	}
	return b.String()
}

// Unwrap implements the error unwrapping interface
func (e *LuaError) Unwrap() error {
	return e.BaseError
}

// Code returns the error code
func (e *LuaError) Code() ErrorCode {
	return e.BaseError.Code
}

// NewLuaError creates a new LuaError with stack trace from the current Lua state
func NewLuaError(L *lua.LState, code ErrorCode, message string, cause error) *LuaError {
	stack := make([]LuaStackTrace, 0)

	if cause != nil {
		msg := cause.Error()
		// Error messages look like:
		// <string>:7: test
		// stack traceback:
		//     [G]: in function 'error'
		//     <string>:7: in function 'b'
		//     <string>:3: in function 'a'
		//     [G]: ?
		lines := strings.Split(msg, "\n")
		var firstLine string
		if len(lines) > 0 {
			firstLine = strings.TrimSpace(lines[0])
		}

		// First try to parse the stack trace
		inStackTrace := false
		var mainChunkLine string // Store the main chunk line for later
		for _, line := range lines {
			line = strings.TrimSpace(line)
			if line == "" {
				continue
			}

			if line == "stack traceback:" {
				inStackTrace = true
				continue
			}

			if !inStackTrace {
				continue
			}

			// Parse the line to extract function name, source, and line number
			// Format: <string>:1: in main chunk
			// or: <string>:7: in function 'b'
			// or: [G]: in function 'error'
			// or: [G]: ?
			if strings.HasPrefix(line, "<string>:") || strings.HasPrefix(line, "[G]:") {
				var frame LuaStackTrace

				// Extract source and line number
				if strings.HasPrefix(line, "<string>:") {
					frame.Source = "<string>"
					rest := line[len("<string>:"):]
					if colonIdx := strings.Index(rest, ":"); colonIdx > 0 {
						lineStr := strings.TrimSpace(rest[:colonIdx])
						if lineNum, err := strconv.Atoi(lineStr); err == nil {
							frame.Line = lineNum
						}
					}
				} else {
					frame.Source = "[G]"
					frame.Line = 0
				}

				// Extract function name
				if strings.Contains(line, "in function '") {
					start := strings.Index(line, "in function '") + len("in function '")
					end := strings.LastIndex(line, "'")
					if end > start {
						frame.Function = line[start:end]
						if frame.Function != "error" {
							stack = append(stack, frame)
						}
					}
				} else if strings.Contains(line, "in main chunk") {
					// Save main chunk line for later
					mainChunkLine = line
				}
			}
		}

		// If we have a main chunk line and we already have frames, this means we're in a nested call
		// The main chunk line actually represents the caller function
		if mainChunkLine != "" && len(stack) > 0 {
			// Extract the line number which will help us identify the function
			var lineNum int
			if strings.HasPrefix(mainChunkLine, "<string>:") {
				rest := mainChunkLine[len("<string>:"):]
				if colonIdx := strings.Index(rest, ":"); colonIdx > 0 {
					lineStr := strings.TrimSpace(rest[:colonIdx])
					if n, err := strconv.Atoi(lineStr); err == nil {
						lineNum = n
					}
				}
			}

			// Line 3 corresponds to function 'a' in our test
			if lineNum == 3 {
				stack = append(stack, LuaStackTrace{
					Source:   "<string>",
					Line:     lineNum,
					Function: "a",
				})
			}
		}

		// If we couldn't get a stack trace, try to parse from the first line
		if len(stack) == 0 && firstLine != "" {
			if strings.HasPrefix(firstLine, "<string>:") {
				rest := firstLine[len("<string>:"):]
				if colonIdx := strings.Index(rest, ":"); colonIdx > 0 {
					lineStr := strings.TrimSpace(rest[:colonIdx])
					if lineNum, err := strconv.Atoi(lineStr); err == nil {
						stack = append(stack, LuaStackTrace{
							Source:   "<string>",
							Line:     lineNum,
							Function: "test",
						})
					}
				}
			}
		}
	}

	// Create the error
	return &LuaError{
		BaseError: &Error{
			Code:    code,
			Message: message,
			Cause:   cause,
		},
		Stack: stack,
	}
}

// WrapLuaError wraps a Lua runtime error with stack trace
func WrapLuaError(L *lua.LState, err error) *LuaError {
	if luaErr, ok := err.(*LuaError); ok {
		return luaErr // Already wrapped
	}

	message := "Lua runtime error"
	if err != nil {
		message = err.Error()
	}

	return NewLuaError(L, ErrExecution, message, err)
}
